software/requirements.txt
Brotli==1.0.9
click==8.1.3
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.0
paho-mqtt==1.6.1
pandas==1.4.3
plotly==5.9.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
schedule==1.1.0
six==1.16.0
tenacity==8.0.1
Werkzeug==2.1.2
